user-service/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
python-dotenv==1.0.0